+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- Expose account external IDs [#582](https://github.com/rokwire/core-building-block/issues/582)
 - Anonymous accounts [#559](https://github.com/rokwire/core-building-block/issues/559)
 - Usernames [#574](https://github.com/rokwire/core-building-block/issues/574)
 - Add release instructions to CONVENTIONS.md [#519](https://github.com/rokwire/core-building-block/issues/519)